+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to the "GitSave" extension will be documented in this file.
 
+## v2.1.0
+
+- Added auto update configuration on change
+- Added additional debug logging
+- Added hide disable message command to Command Palette
+
 ## v2.0.0
 
 - Revert back to absolute path